refactor: Divide course-attempts.page.tsx by areas of responsibility
